requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
boto3==1.17.45
botocore==1.20.45
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-environ==0.4.5
django-heroku==0.3.1
filelock==3.0.12
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
jmespath==0.10.0
msgpack==1.0.2
numpy==1.19.4
pandas==1.2.0
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.0
pyOpenSSL==20.0.1
pyperclip==1.8.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
s3transfer==0.3.6
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
twilio==6.50.1
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.2
virtualenv==20.4.3
virtualenv-clone==0.5.4
whitenoise==5.2.0
zope.interface==5.3.0
